Edits

+++ Functions/Recursive/FactorialOfNNumbers.c
@@ -1,24 +1,24 @@
 #include <stdio.h>
 
 
-int smallfactorial(int n)
+long int smallfactorial(int n)
 {
 	return (n==1 ? 1 : n*smallfactorial(n-1));
 }
 
 int main()
 {
-	int n;
+	long int n;
 	printf("Enter Number of Nos for which you want factorial: \n");
-	scanf("%d", &n);
+	scanf("%ld", &n);
 	
 	for(int i = 1; i <= n; i++)
 	{
-		int fact;
+		long int fact;
 		
 		fact = smallfactorial(i);
 		
-		printf("Factorial value  of %d = %d\n", i, fact);
+		printf("Factorial value  of %d = %ld\n", i, fact);
 	}	
 	
 return 0;
